+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [Unreleased]
+
+### Fixed
+- Validation for an unaccessible resource inside a bundle ([rhbz#1462248])
+
+[rhbz#1462248]: https://bugzilla.redhat.com/show_bug.cgi?id=1462248
+
+
 ## [0.9.165] - 2018-06-21
 
 ### Added